Upgrade dependencies, style menu

+++ src/lib/dateToLocaleString.ts
@@ -1,7 +1,7 @@
 const dateToLocaleString = (date: Date): string => {
   const options = {
     year: 'numeric' as const,
-    month: 'long' as const,
+    month: 'short' as const,
   };
 
   return date.toLocaleDateString('en-US', options);
